Consolidate relcache cache invalidation cleanup
Register a single pg_sorted_heap relcache dispatcher instead of separate pg_sorted_heap and sorted_heap callbacks. The dispatcher keeps the invalidation domains explicit: insert-side directed-placement zone maps, HNSW L0/upper session caches, and sorted_heap Table AM relation metadata.

Factor the repeated pg_sorted_heap zone-map hash lifecycle into helpers for creating the relation hash, freeing one relation entry, and destroying all entries. This removes duplicated hash_seq/hash_destroy blocks from global relcache invalidation and the overflow guard without changing cache semantics.

Factor sorted_heap relation zone-map overflow cleanup into sorted_heap_relinfo_reset_zonemap_cache() and reuse it for relcache and DDL invalidation paths. Also cast sorted_hnsw scan-stat uint64 counters for PRIu64 formatting so PG16 builds stay warning-clean.

Verification: make test-release PG_CONFIG=/opt/homebrew/opt/postgresql@18/bin/pg_config TMP_SELFTEST_ROOT=/private/tmp passed; clean build/install/installcheck passed for PG16, PG17, and PG18 with REGRESS='pg_sorted_heap sorted_hnsw graph_rag'; git diff --check passed.

+++ src/sorted_hnsw.c
@@ -598,11 +598,11 @@ sorted_hnsw_scan_stats(PG_FUNCTION_ARGS)
 		" last={ef=%d,nodes=%d,l0_candidates=%d,initial_results=%d,"
 		"topup_ef=%d,topup_candidates=%d,topup_results=%d,"
 		"fallback_results=%d,final_results=%d,exact_fallback=%s}",
-		shnsw_scan_stats.calls,
-		shnsw_scan_stats.l0_searches,
-		shnsw_scan_stats.topup_searches,
-		shnsw_scan_stats.exact_fallbacks,
-		shnsw_scan_stats.exact_fallback_wins,
+		(unsigned long long) shnsw_scan_stats.calls,
+		(unsigned long long) shnsw_scan_stats.l0_searches,
+		(unsigned long long) shnsw_scan_stats.topup_searches,
+		(unsigned long long) shnsw_scan_stats.exact_fallbacks,
+		(unsigned long long) shnsw_scan_stats.exact_fallback_wins,
 		shnsw_scan_stats.last_ef,
 		shnsw_scan_stats.last_nodes,
 		shnsw_scan_stats.last_l0_candidates,
